+++ ompi/mca/op/sve/op_sve_component.c
@@ -26,8 +26,8 @@
 #include "ompi/op/op.h"
 #include "ompi/mca/op/op.h"
 #include "ompi/mca/op/base/base.h"
-#include "ompi/mca/op/arm_sve_op/op_sve.h"
-#include "ompi/mca/op/arm_sve_op/op_sve_functions.h"
+#include "ompi/mca/op/sve/op_sve.h"
+#include "ompi/mca/op/sve/op_sve_functions.h"
 
 static int sve_component_open(void);
 static int sve_component_close(void);
